Refactor|Model Renderer|libgui: Added ModelDrawable::Appearance
Added a new object that contains parameters that affect how a model
gets drawn. ModelDrawable::Appearance contains the rendering passes,
mask for active passes, material to use for each pass, and callbacks
for changing the program and pass.

This allows StateAnimator to maintain this information internally for
each animated object, rather than having to pass it to the model
renderer that does the draw calls.