libgui: Added MultiAtlas: expanding collection of atlases
One atlas cannot hold an infinite number of images, but a MultiAtlas
can.

Atlas was modified to allow making allocations tentatively in
deferred mode, so that MultiAtlas can see which atlas can hold all
the allocations in a group.

The atlas allocators were slightly modified to allow specifying a
pre-existing Id for new allocations.