Fixed|Refactor|Thinker|Hexen: Mobj-on-mobj checks shouldn't duplicate the mobj
When performing a check of whether a mobj was resting on another mobj,
Hexen would duplicate the entire thinker (!), with the objective of
restoring the object's position after the check.

As a side effect, the internal thinker data was being reinitialized
after the duplication, causing 3D model state to revert back to its
defaults.

Now the saving and restoring only affects the object's position and
momentum, as appropriate.