Refactor|World|Client: Reacting to mobj damage
The client is now notified of damage to mobjs via MobjThinkerData.
It can then spawn particle generators and check if the mobj's model
asset has a script callback for reacting to damage.

Also renamed the script variables in StateAnimator with ALL CAPS to
reflect that they are intended to be constants.