Fixed|Renderer: Configuring all fog parameters from Map Info
In Hexen, the fog on a map is being enabled with a fade table
called "FOGMAP". However, the fog start/end and other parameters were
not being set at all in this case.

The logic was modified to always configure the fog parameters, but
then disable the fog if it should not be used on the map.