Terminated branch audio-cleanup
It has become increasingly clear that the architecture suggested by
libgui's audio::Sound is unworkable due to the design of the mechanism
for looped playback (ambient sounds). An altogether different approach
(which, largely follows vanilla) would achieve an objectively better
solution.

My new plan of action is to focus on ambient sounds, starting with the
unification and commonization of the existing mechanisms for this in
Heretic and Hexen.

Although, it may be possible to adapt some of this work into a future
branch, at a later stage (e.g., new mechanisms for managing drivers
and playback channels).