Log|libcore|libdoomsday: Interactive script output bypasses filtering
When the user enters commands in the console, and the commands produce
output, it is better to allow these entries to bypass the chosen
filters. Log entries created during execution of console commands
whose source is either "console" or "cmdline" are now flagged as
Interactive, and LogFilter allows these to pass in the Script domain.

This is not a perfect solution, though, because not all console
commands (yet) use SCR when printing output.

IssueID #2074