Cleanup|libdoomsday: Renamed res::System to Resources
Continuing the de-System-ification of the classes. Subsystems should
be named according to their domain: World, Resources, etc. Also, the
primary class of a domain should not be in the domain's namespace,
but in the top-level/global namespace.

(FileSystem and WindowSystem have "system" in the name because those
are the commonly used terms for the domains.)