Resources|libdoomsday: Identifying data files and indexing WAD files
DoomsdayApp checks all the configured and known locations for WAD files
and populates the /local/wads folder. Currently this is done in
addition to the client's FS1-based code for doing the same task
(that is slated to be retired later).

DataBundle checks the indexed data files (WADs and other formats)
against the databundles.dei registry, which defines the criteria for
generating packages out of known files.

res::LumpDirectory is a new class in libdoomsday for parsing a WAD
file's lump directory and performing basic lookup operations in it.