Resources|libdoomsday: Startup files are loaded as FS2 packages
The old game startup resource locator code was disabled as obsolete.
Recognized data files (i.e., data bundles) are already indexed in the
FS2 index, and the recognized packages have been linked under
/sys/bundles. This makes it fast and easy to check which required
files are present.

The data files themselves, and any additional user-provided files, are
still loaded using the old FS1 based resource code. Loading a
data bundle as a package currently does nothing except mark the
package as loaded. In the future, this should be further integrated
so that when a data bundle is loaded, the data is instantly taken
into use.