UI|Client: Removed first startup busy mode, added a splash window
The first startup busy mode was mostly used when waiting for the
game startup resources to be located. With that step removed, the
busy mode would not take long enough to warrant showing a progress
animation.

Todo for later: Instead of a non-GL splash screen, show the
ClientWindow with an initial small widget tree that doesn't take
long to initialize. The startup operations should be done via a
TaskPool instead of busy mode.