Resources|libdoomsday: Read Snowberry metadata, nested data files, autoloading
Read the Snowberry metadata (if available) when generating packages out
of classic data files and resource packs. In case the folder structure
implies autoloading, register the appropriate load dependencies into
the packages. This will ultimately make the old FS1 autoload code
obsolete.

If metadata is not available, the file names are checked for suffixes
that look like version numbers.