Refactor|libcore|Client: Improved Version class API
The default constructor now creates a null version and there is a
static method for getting info about the current build. This is
more readable than the previous implicit behavior of Version().

This makes it also less of a performance hit in situations where
Version instances get constructed often. (No calls to Time().)