libappfw: Child widget virtualization for large data sets
ChildWidgetOrganizer can now be told to use widgets only for the
visible region of the screen. The rest of the items are handled by
estimation.

Currently this only works for vertical single-column lists.

MenuWidget supports virtualization.

Todo: Apply some finesse in handling the strut length calculations
with regard to to items becoming ready for drawing.