Refactor|libcore: Audience members are automatically removed upon deletion
Each audience member now has a link back to the audience, so that
the observer relationship can be terminated by either party on the
event of deletion of either the audience or the observer.

This makes it unnecessary to remove audience members manually
before they are deleted, increasing the robustness of the system.