Refactor|Resources: Moved texture resources to libdoomsday (big commit)
This commit splits Texture to the generic res::Texture and
client-specific ClientTexture. Other classes moved to libdoomsday
include TextureScheme, TextureManifest, PatchName, and CompositeTexture.

All the texture-related objects are now owned by the Textures class,
which in turn owned by libdoomsday's Resources.

Also fixed a couple of errors in the implementation:

- Retaining references to TextureVariantSpec without any kind of
  reference counting, leading to use-after-free errors during engine
  reset.
- Pointers to owning TextureSchemes are now kept as direct pointers,
  which avoids problems during teardown (trying to iterate partially
  deleted scheme lists).