Refactor: Moved materials to libdoomsday's world domain (big commit)
As they are currently implemented and used, the materials are the best
fit for the world domain. The classes are now all in the `world`
namespace (save for the client-specific ones).

The supporting MapElement and DmuArgs classes were also moved to
libdoomsday.

Because libdoomsday is used by the client, server, and all plugins,
this makes materials now available everywhere. An opaque pointer type
is still used in the published runtime APIs for C language support.

The MaterialArchive runtime API was removed as unnecessary, as
the class can now be directly accessed.