UI|Home|Client: Minimizing the loaded game to show the Home UI
The Home UI can now be shown while a game is running, allowing
access to game switching, savegames, and multiplayer games.
Previously these were available via popup dialogs, but now one can
use Home as usual, and the game remains playable in a minimized size
in the bottom of the screen.