Input|Client|Refactor: Xbox One preset added; axis scaling factor default is 1.0
The SDL range of 10000 units should not be exposed to the user in
the form of the scale factor being 1/10000. Instead, apply the SDL
scaling internally and have a 1.0 scaling factor as the cvar.

Added a new axis parameter for offsetting the values.

The _controllers.de_ script now has bindings and configuration preset
for the Xbox One controller. The PS3 buttons preset was slightly
tweaked.

Renamed input device classes according to preferred conventions.