Fixed|Input: Hat bindings for different angles; added "inspectinputdevice"
When creating joystick hat event bindings, the specified hat angle
was ignored when checking for overridden bindings. This meant that only
the latest hat binding was ever in effect.

If the device contains multiple hats, up to four hats are checked
and events are generated appropriately.

The "inspectinputdevice" command can be used to check the description
(and status) of a single input device.

Fixes #2171