CMake|Builder: Changed default value for the Turbo build option
The idea is to only enable Turbo for release builds, because when
debugging it is more valuable to use the original source files for
breakpoints, etc.

However, in multiconfiguration projects like MSVC and Xcode, CMake
does not support different sets of source files per configuration,
so Turbo is always disabled by default and must be enabled manually.

(Autobuilder/CI uses Turbo for all builds.)