Refactor|GL|libgui: Use Qt 5.4 OpenGL APIs, renamed GLTarget/GLFramebuffer
Qt 5 comes with new OpenGL classes (prefix "QOpenGL") and a portable
way to access OpenGL and extension functions. This commit takes these
into use and removes the old custom "glentrypoints".

The functionality of the GL code was not changed, and the OpenGL
version is kept at 2.1 (for now). Bumping to OpenGL 3/4 will require
larger revisions in the world renderer.

Multisampling is temporarily disabled pending a better implementation.

`GLTarget` was renamed to `GLFramebuffer` to more clearly indicate that
this is a wrapper around an FBO.

(The old) `GLFramebuffer` was renamed to `GLTextureFramebuffer` and is
now a subclass of (the new) `GLFramebuffer`, to indicate that this is
a specialized FBO that owns the buffer textures.