Refactor|Cleanup|libgui: Cleaning up Canvas/CanvasWindow and window init/drawing
Cutting down on the cruft. The Qt 5 OpenGL APIs allow for a cleaner
implementation of window initialization and refresh. The "frame
swapped" signal is particularly helpful.

Using an audience for drawing the window contents is not the best
idea because the order of audience callbacks is unspecified. Now
the draw() method is a simple virtual method call back to the parent
CanvasWindow.

GLInfo now checks for the EXT_timer_query extension.