Refactor|Renderer|UI: Post-processing shader applied during view composition
Instead of having fx::PostProcessing as a lens effect, ViewCompositor
now applies the post-processing shader when it blits the game view to
the window as one of the layers of the view.

This has the benefit that the post-processor doesn't need a separate
framebuffer of its own because it can use the existing game view
texture and draw directly in the window framebuffer.

fx::Resize was also removed because the game view resolution can be
trivially changed by resizing the game view texture.