Refactor|Cleanup|Renderer: Less complicated view rendering
This commit fixes issues when the game widget does not cover the
entire window.

To summarize, the viewport no longer needs modifying when a frame
is being drawn: it's either the player's game view texture (that
contains only a single full view), or ViewCompositor is drawing the
view layers within the appropriate rectangle in the window and
handles the viewport on its own.