Resources|FS|libdoomsday: Loading data files via packages; added "defsPath"
When a package is loaded, the corresponding FS1 data file is loaded
by libdoomsday's Resources class. The res::DoomsdayPackage utility
provides the URI for loading.

Packages may now specify a "defsPath" in their metadata. This is
assumed to contain DED files that will automatically be loaded when
definitions are read.

IssueID #2178