Fixed|GL|libgui: Fixed alternative GLFramebuffer attachments
The alternative attachments now work with both textures and render
buffers. This is necessary because the rendering target is a render
buffer when FSAA is enabled, but is a texture otherwise.

The player 3D weapons use a temporary depth/stencil render buffer that
is separate from the rest of the frame (previously was a texture).