Shell|libshell: Improvements for opening connection to a server
A shell::Link now needs to be opened with a separate method so that
the object can be fully set up before starting a connection.

The GUI Shell app now connects to new local servers by waiting for the
first beacon broadcast that indicates the server is up and running.
Blindly trying to open the socket too early seems to trigger a
lower-level timeout that takes far longer for the connection to be
established.