libgamefw|Map Data: Common map spawn spot flags
The new GFW_MAPSPOT_* flags are currently stored in the MapSpotFlags
property of the entity database.

Todo: Some game-specific flags are missing. All flags should be present
so that full translation is possible in all games.

Todo: Replace the old Flags property; the “importidtech1” plugin needs
most of the work.