Refactor|Model Renderer: Improved model animator state serialization
The special ASSET member of the animator state was renamed to
“__asset__” so it will be ignored in serialization.

The special THING member was also renamed to “__thing__”, and a
method called “thing()” was added to the Render.StateAnimator class
for convenience.

For scripting purposes, the model state animator now uses
AnimationValues for shader and transformation variables. These can
be (de)serialized as any other value.

The previous use of NativePointerValue was removed.