libcommon|libdoomsday: Thinker ID mapping for deserializing internal map state
When reading internal map state, it is crucial to know which internal
state belongs to which thinker. However, the challenge is that when
thinkers are created, they don’t yet have ThinkerData and therefore
no private IDs either. The thinker ID mapping allows restoring the
old IDs.

The save version was bumped so that the private thinker IDs can be
included in the save data.