FS|Performance|libcore: Cache ZIP metadata
ZIP directory metadata is now cached so that we can populate the
filesystem without having to actually read the ZIP file contents.

The cached metadata is stored in “/home/cache/metadata/ZipArchive”.

de::ZipArchive was modified to allow reading the central directory
either from the original file or from the cached data.