Fixed|UI|Home: Random crash after editing package profiles
Opening a PackagesDialog would refresh all packages, which would
result in a game readiness update notification, which in turn would
clear the game column’s contents, leading to a crash because the user
was in the process of editing a game profile.

Now game columns do not clear their contents when an update is needed,
but instead add, remove, or update individual items.