Fixed|Model Renderer: Repeated triggering of a single thing state
If a thing stays in a single-state loop (e.g., HEAD_STND), don’t keep
retriggering the state. Sequence triggering should only occur if
the thing state changes or when a model animation sequence ends.

Warn if a model definition tries to use animations on a model that
has none.