Fixed|Client|libcore|libappfw: More graceful fatal errors
A fatal error should not crash the engine, just pop up an error message.

Client's BusyRunner was ignoring the busy worker abort notification.
Now it causes the busy worker to be instantly killed.

While a fatal error message is displayed the main game is window
is just hidden, not deleted. Deleting everything while the objects
are in use is not a great idea.

IssueID #2251