Audio: Added the GeneralUser GS SoundFont to net.dengine.client
The client now has a default MIDI SoundFont included in its data
package so it is possible to play the vanilla MIDI music
out of the box with a known level of quality.

See the included GeneralUser GS readme and license for more
information.

Source: http://www.schristiancollins.com/generaluser.php