All Games|Client|SaveGame: Restore object links via a new mechanism
After a save has been loaded, object cross-references haven't always
been correctly restored to what they were when the save was made.

The previous method of validating the cross-references was extended
to actually restore all cross-references to point to the correct
mobjs using the private thinker IDs.

While this obsoletes the cross-reference restoration done by
MapStateReaders, the old code still is necessary when loading saves
that don't contain the ObjectState info file.

Also noteworthy is that private IDs are persistent across a save
and restore.

IssueID #2261