Refactor|libdoomsday|Plugins: Game exports queried individually
Instead of GetGameAPI returning a struct with the exported game
functions, GetGameAPI now takes the name of the function and returns
a pointer to that particular function.

This is more flexible and future-proof, and allows using C++ types in
the arguments and return values. The exports from libcommon are
also handled in one place.