FS|Network: Improving remote file access
RemoteFile now stores the actual file in the cache using the
original name, and is derived from LinkFile to provide access to it.
This is similar to how local data files are linked as packages.

The special path "/sys/server/files" is where servers are expected
to keep files for remote access.