FS|Resources|Refactor: Querying packages from remote repositories
A repository Link now has knowledge of which packages are available
remotely, and provides the list of remote files that correspond
the requested packages, so that they can be downloaded. The Link
will also manage the root folder representing the repository locally.

IdgamesLink parses the file tree index and generates a package index
based on it.