Fixed|FS|Multiplayer: Loading data files; checking server's packages
The "dataFiles" in package metadata are now expected to be relative
to the package root, but some packages are just links to a WAD, etc.
In these cases, "dataFiles" is ignored.

The initial population of the server's public folder was sometimes
not done soon enough, because the notification about when the
connection is ready was sent first to PackageDownloader. Now the
downloader defers its reaction so RemoteFeeds can begin population
first.