Widgets|libappfw: UI style reloading
If the DPI changes, the UI style needs to be updated. The length rules use a common DPI factor so they don’t need to be reloaded, but fonts and other assets will need to be refreshed.

Widgets update their own style parameters.

Work in progress: this does not work completely yet.