Performance: Manipulating and copying ServerInfos
The Record-based shell::ServerInfo was being copied multiple times continuously during discovery and update of servers. Apply a copy-on-write mechanism internally in ServerInfo so that it can be used more efficiently.

Also Address now caches localhost check results and the text representation of the address, so these don’t have to be done more than once.

IssueID #2285