Fixed|Music: Defined external music file ignored
The way Hexen loads the MAPINFO and SNDINFO lumps was causing custom music definitions to be ignored.

The MAPINFO translator (that generates corresponding DEDs) was adding music definitions that would be used even when custom music was loaded. Now these generated definitions are only present when parsing custom MAPINFO data. (Non-custom map infos are expected to have a music definition available matching the map ID.)

In Hexen, SNDINFO is applied by default so that if a matching music definition is available (by ID), the map info is changed to use that music (that may be a custom music definition).

IssueID #2082