Gameplay|All Games: Compatibility option for pushable momentum
When pushing objects with the MF2_PUSHABLE flag, the previous (vanillla?) behavior is that the pushed objects may accumulate unlimited amounts of momentum.

The new compatibility cvar “game-objects-pushable-limit” (default 1) causes the pushed objects’ momentum to not exceed the speed of the object doing the pushing. This makes the behavior more physically accurate.