Removed plugin functionality; GameKit provides games
Dynamic plugins are not needed. Instead, games are linked to GameKit, which registers them as extensions.

Games are now built as private dynamic libraries linked to GameKit. Ideally they would be static libraries, but dynamic ones allow duplicate symbols to be hidden.

GameFw is a separate library providing common functionality for games.