Broke the world apart
ClientServerWorld is now split to ClientWorld and ServerWorld, and the common parts are in the libdoomsday base class world::World.

The backend for MPE (MapBuilder) was moved to libdoomsday so it can be used for converting the map to load.

World.Thing.startSound() is now a pure script function, making a call to Audio.startSound(). This simplifies the bindings setup in libdoomsday, where the audio subsystem is not (yet) available.