Heretic|Intermission: Using a patch image as background
The Intermission Background of a Map Info can now be set to a patch in Heretic.

Instead of attempting to solve this at a lower level, the intermission drawing routine now checks whether the defined image is a flat or a patch, and draws it with bespoke code for the type of resource.

IssueID #2427