Fixed|Heretic: Sky textures are incorrectly sized
The SKY* patches in Heretic are 200 pixels tall even though the texture is declared as 128 pixels tall. The extra height is supposed to make the sky extend upward to facilitate looking up. However, for some reason only Composite::dimensions() is updated to account for this extended height, and not Composite::logicalDimensions(). If both are updated, skyfix walls appear as black in DOOM (for an unknown reason). Therefore, apply a hacky workaround that uses the true composite dimensions for sky textures only.

IssueID #2446